Struct budvm::ir::CodeBlockBuilder

pub struct CodeBlockBuilder { /* private fields */ }

A type that helps build CodeBlocks.

Implementations

impl CodeBlockBuilder

pub fn new_argument(&mut self, name: impl Into<Symbol>) -> Argument

Adds a new argument with the given name.

pub fn argument(&self, name: &Symbol) -> Option<Argument>

Finds an argument by a given name, or None if not found.

pub fn new_label(&mut self) -> Label

Create a new label.

pub fn named_label(&mut self, name: impl Into<Symbol>) -> Label

Create a new label with a given name.

pub fn push(&mut self, operation: Instruction)

Push an instruction.

pub fn label(&mut self, label: Label)

Label the next instruction as label.

pub fn lookup(&self, symbol: &Symbol) -> Option<&ScopeSymbol>

Looks up a symbol.

pub fn loop_info(&self, name: Option<&Symbol>) -> Option<&LoopInfo>

Returns the loop information for a given name, or the current loop if no name is provided.

pub fn store_into_destination(
&mut self,
value: LiteralOrSource,
destination: Destination
)

Adds the appropriate instruction to store value into destination.

pub fn load_from_symbol(
&mut self,
symbol: &Symbol,
destination: Destination
) -> Result<(), LinkError>

Adds the correct instruction to load a value from symbol and store it in destination.

pub fn variable_index_from_name(&mut self, name: &Symbol) -> Variable

Looks up an existing location for a variable with the provided name. If an existing location is not found, new space will be allocated for it and returned.

pub fn new_temporary_variable(&mut self) -> Variable

Creates a new temporary variable.

Internally this simply uses a counter to create a new variable each time this is called named $1, $2, and so on.

pub fn finish(self) -> CodeBlock

Returns the completed code block.

pub fn begin_loop(
&mut self,
name: Option<Symbol>,
result: Destination
) -> LoopScope<'_, Self>

Begins a loop with the given name. The result of the loop will be stored in result. If the loop does not return a result, the destination will be untouched.

pub fn variables(&self) -> &HashMap<Symbol, Variable>

Returns the collection of known variables.