Struct budvm::ir::LoopScope

pub struct LoopScope<'a, T>where
    T: BorrowMut<CodeBlockBuilder>,{
    pub break_label: Label,
    pub continue_label: Label,
    /* private fields */
}

A loop within a CodeBlockBuilder.

Fields
break_label: Label

The label for a break operation.

continue_label: Label

The label for a continue operation.

Implementations

impl<'a, T> LoopScope<'a, T>where
T: BorrowMut<CodeBlockBuilder>,

pub fn label_break(&mut self)

Marks the next instruction as where the break operation should jump to.

pub fn label_continue(&mut self)

Marks the next instruction as where the continue operation should jump to.