Trait budvm::ir::Scope

pub trait Scope {
    type Environment: Environment;

    // Required methods
    fn resolve_function_vtable_index(&self, name: &Symbol) -> Option<usize>;
    fn map_each_symbol(&self, callback: &mut impl FnMut(Symbol, ScopeSymbolKind));
    fn define_function(
        &mut self,
        function: Function<<Self::Environment as Environment>::Intrinsic>
    ) -> Option<usize>;
    fn define_persistent_variable(&mut self, name: Symbol, variable: Variable);
}

A scope for linking and compiling code.

Required Associated Types

type Environment: Environment

The environment for this scope.

This is used when instantiating literals as values.

Required Methods

fn resolve_function_vtable_index(&self, name: &Symbol) -> Option<usize>

Returns the vtable index of a function with the provided name.

fn map_each_symbol(&self, callback: &mut impl FnMut(Symbol, ScopeSymbolKind))

Invokes callback for each symbol defined in this scope.

fn define_function( &mut self, function: Function<<Self::Environment as Environment>::Intrinsic> ) -> Option<usize>

Defines a function with the provided name.

fn define_persistent_variable(&mut self, name: Symbol, variable: Variable)

Defines a persistent variable.

This is used to enable interactive sessions.