Module budvm::ir

An intermediate representation of virtual machine instructions.

This intermediate representation provides conveniences for handling variables, labels, and function calls.

Structs

CodeBlock

A block of intermediate instructions.

CodeBlockBuilder

A type that helps build CodeBlocks.

Function

A function, in its intermediate form.

Label

A label that can be jumped to.

LoopInfo

Information about a loop.

LoopScope

A loop within a CodeBlockBuilder.

Module

A collection of functions and modules.

Variable

A reference to a local variable.

Enums

CompareAction

An action to take during an Instruction::Compare.

Destination

A destination for a value.

Instruction

An intermediate representation of an crate::Instruction.

LinkError

An error occurred while linking code.

Literal

A literal value.

LiteralOrSource

A literal value or a location of a value

ScopeSymbol

A registered symbol.

ScopeSymbolKind

The kind of a ScopeSymbol.

ValueSource

The source of a value.

Traits

Scope

A scope for linking and compiling code.