Enum budvm::ir::Instruction

pub enum Instruction {

    Add {
        left: LiteralOrSource,
        right: LiteralOrSource,
        destination: Destination,
    },
    Sub {
        left: LiteralOrSource,
        right: LiteralOrSource,
        destination: Destination,
    },
    Multiply {
        left: LiteralOrSource,
        right: LiteralOrSource,
        destination: Destination,
    },
    Divide {
        left: LiteralOrSource,
        right: LiteralOrSource,
        destination: Destination,
    },
    And {
        left: LiteralOrSource,
        right: LiteralOrSource,
        destination: Destination,
    },
    Or {
        left: LiteralOrSource,
        right: LiteralOrSource,
        destination: Destination,
    },
    Xor {
        left: LiteralOrSource,
        right: LiteralOrSource,
        destination: Destination,
    },
    Not {
        value: LiteralOrSource,
        destination: Destination,
    },
    If {
        condition: LiteralOrSource,
        false_jump_to: Label,
    },
    JumpTo(Label),
    Label(Label),
    Compare {
        comparison: Comparison,
        left: LiteralOrSource,
        right: LiteralOrSource,
        action: CompareAction,
    },
    Push(LiteralOrSource),
    Load {
        value: LiteralOrSource,
        variable: Variable,
    },
    Return(Option<LiteralOrSource>),
    Call {
        function: Option<Symbol>,
        arg_count: usize,
        destination: Destination,
    },
    CallIntrinsic {
        intrinsic: Intrinsic,
        arg_count: usize,
        destination: Destination,
    },
    CallInstance {
        target: Option<LiteralOrSource>,
        name: Symbol,
        arg_count: usize,
        destination: Destination,
    },

}

An intermediate representation of an crate::Instruction.

Variants

Add

Fields

left: LiteralOrSource

The left hand side of the operation.

right: LiteralOrSource

The right hand side of the operation.

destination: Destination

The destination for the result to be stored in.

Adds left and right and places the result in destination.

If this operation causes an overflow, Value::Void will be stored in the destination instead.

Sub

Fields

left: LiteralOrSource

The left hand side of the operation.

right: LiteralOrSource

The right hand side of the operation.

destination: Destination

The destination for the result to be stored in.

Subtracts right from left and places the result in destination.

If this operation causes an overflow, Value::Void will be stored in the destination instead.

Multiply

Fields

left: LiteralOrSource

The left hand side of the operation.

right: LiteralOrSource

The right hand side of the operation.

destination: Destination

The destination for the result to be stored in.

Multiply left by right and places the result in destination.

If this operation causes an overflow, Value::Void will be stored in the destination instead.

Divide

Fields

left: LiteralOrSource

The left hand side of the operation.

right: LiteralOrSource

The right hand side of the operation.

destination: Destination

The destination for the result to be stored in.

Divides left by right and places the result in destination.

If this operation causes an overflow, Value::Void will be stored in the destination instead.

And

Fields

left: LiteralOrSource

The left hand side of the operation.

right: LiteralOrSource

The right hand side of the operation.

destination: Destination

The destination for the result to be stored in.

Performs an and operation with left and right, storing the result in destination.

If both operands are integers, this performs a bitwise operation. Dynamic types can also implement custom behaviors for this operation.

If no other implementations are provided for the given types, the result will be a boolean evaluation of left.is_truthy() and right.is_truthy().

Or

Fields

left: LiteralOrSource

The left hand side of the operation.

right: LiteralOrSource

The right hand side of the operation.

destination: Destination

The destination for the result to be stored in.

Performs an or operation with left and right, storing the result in destination.

If both operands are integers, this performs a bitwise operation. Dynamic types can also implement custom behaviors for this operation.

If no other implementations are provided for the given types, the result will be a boolean evaluation of left.is_truthy() or right.is_truthy().

Xor

Fields

left: LiteralOrSource

The left hand side of the operation.

right: LiteralOrSource

The right hand side of the operation.

destination: Destination

The destination for the result to be stored in.

Performs a xor operation with left and right, storing the result in destination.

If both operands are integers, this performs a bitwise operation. Dynamic types can also implement custom behaviors for this operation.

If no other implementations are provided for the given types, the result will be a boolean evaluation of left.is_truthy() xor right.is_truthy().

Not

Fields

value: LiteralOrSource

The left hand side of the operation.

destination: Destination

The destination for the result to be stored in.

Performs a not operation for value, storing the result in destination.

If the operands is an integer, this performs a bitwise operation. Dynamic types can also implement custom behaviors for this operation.

If no other implementations are provided for the given types, the result will be a boolean evaluation of not value.is_truthy().

If

Fields

condition: LiteralOrSource

The source of the condition.

false_jump_to: Label

The label to jump to if the condition is falsey.

Checks condition.is_truthy(), jumping to the target label if false.

If truthy, the virtual machine continues executing the next instruction in sequence.

If not truthy, the virtual machine jumps to label false_jump_to.

JumpTo(Label)

Jump execution to the address of the given label.

Label(Label)

Labels the next instruction with the given Label.

Compare

Fields

comparison: Comparison

The comparison to perform.

left: LiteralOrSource

The left hand side of the operation.

right: LiteralOrSource

The right hand side of the operation.

action: CompareAction

The action to take with the result of the comparison.

Compares left and right using comparison to evaluate a boolean result.

If CompareAction::Store is used, the boolean result will be stored in the provided destination.

If CompareAction::JumpIfFalse is used and the result is false, execution will jump to the label specified. If the result is true, the next instruction will continue executing.

Push(LiteralOrSource)

Pushes a value to the stack.

Load

Fields

value: LiteralOrSource

The index of the variable to store the value in.

variable: Variable

The value or source of the value to store.

Loads a value into a variable.

Return(Option<LiteralOrSource>)

Returns from the current stack frame.

Call

Fields

function: Option<Symbol>

The name of the function to call. If None, the current function is called recursively.

arg_count: usize

The number of arguments on the stack that should be used as arguments to this call.

destination: Destination

The destination for the result of the call.

Calls a function.

When calling a function, values on the stack are “passed” to the function being pushed to the stack before calling the function. To ensure the correct number of arguments are taken even when variable argument lists are supported, the number of arguments is passed and controls the baseline of the stack.

Upon returning from a function call, the arguments will no longer be on the stack. The value returned from the function (or Value::Void if no value was returned) will be placed in destination.

CallIntrinsic

Fields

intrinsic: Intrinsic

The runtime intrinsic to call.

arg_count: usize

The number of arguments on the stack that should be used as arguments to this call.

destination: Destination

The destination for the result of the call.

Calls an intrinsic runtime function.

When calling a function, values on the stack are “passed” to the function being pushed to the stack before calling the function. To ensure the correct number of arguments are taken even when variable argument lists are supported, the number of arguments is passed and controls the baseline of the stack.

Upon returning from a function call, the arguments will no longer be on the stack. The value returned from the function (or Value::Void if no value was returned) will be placed in destination.

CallInstance

Fields

target: Option<LiteralOrSource>

The target of the function call. If None, the value on the stack prior to the arguments is the target of the call.

name: Symbol

The name of the function to call.

arg_count: usize

The number of arguments on the stack that should be used as arguments to this call.

destination: Destination

The destination for the result of the call.

Calls a function by name on a value.

When calling a function, values on the stack are “passed” to the function being pushed to the stack before calling the function. To ensure the correct number of arguments are taken even when variable argument lists are supported, the number of arguments is passed and controls the baseline of the stack.

Upon returning from a function call, the arguments will no longer be on the stack. The value returned from the function (or Value::Void if no value was returned) will be placed in destination.