Struct budvm::Stack

pub struct Stack { /* private fields */ }

A stack of Values.

Implementations

impl Stack

pub fn new(initial_capacity: usize, maximum_capacity: usize) -> Self

Returns a new stack with enough reserved space to store initial_capacity values without reallocating and will not allow pushing more than maximum_capacity values.

pub fn push(&mut self, value: Value) -> Result<(), FaultKind>

Pushes value to the stack.

Errors

Returns FaultKind::StackOverflow if the stack’s maximum capacity has been reached.

pub fn extend<Args, ArgsIter>(&mut self, args: Args) -> Result<usize, FaultKind>where
Args: IntoIterator<Item = Value, IntoIter = ArgsIter>,
ArgsIter: Iterator<Item = Value> + ExactSizeIterator + DoubleEndedIterator,

Pushes multiple arguments to the stack.

pub fn pop_n(&mut self, count: usize) -> PoppedValues<'_>

Pops count elements from the top of the stack.

This iterator returns the values in the sequence that they are ordered on the stack, which is different than calling pop() count times sequentially. For example, if the stack contains [0, 1, 2, 3], calling pop() twice will result in 3, 2. Calling pop_n(2) will result in 2, 3.

pub fn top(&self) -> Result<&Value, FaultKind>

Returns a reference to the top Value on the stack, or returns a FaultKind::StackUnderflow if no values are present.

pub fn top_mut(&mut self) -> Result<&mut Value, FaultKind>

Returns a reference to the top Value on the stack, or returns a FaultKind::StackUnderflow if no values are present.

pub fn pop(&mut self) -> Result<Value, FaultKind>

Pops a Value from the stack.

Errors

Returns FaultKind::StackUnderflow if the stack is empty.

pub fn truncate(&mut self, new_length: usize)

Truncates the stack to new_length. If the new_length is larger than the current length, this function does nothing.

pub fn len(&self) -> usize

Returns the number of Values contained in this stack.

pub fn is_empty(&self) -> bool

Returns true if this stack has no values.

pub const fn remaining_capacity(&self) -> usize

Returns the number of Values that can be pushed to this stack before a FaultKind::StackOverflow is raised.

pub fn grow_by(&mut self, additional_voids: usize) -> Result<(), FaultKind>

Grows the stack by additional_voids, inserting Value::Void to fill in any newly allocated space.